CVE-2024-11053.json: adjusted CVSS scoring: MEDIUM 5.3
The security problem this describes is mostly a risk that a user can
accidentally stumble upon this. It is VERY hard for an attacker to
exploit.

- "vectorString": "CVSS:3.1/AV:L/AC:H/PR:L/UI:N/S:U/C:L/I:H/A:N",

I'll explain my reasoning:

- "attackVector": "LOCAL",

Because an attacker needs acess to a user's local .netrc file in order
to perform an attack. It could be a NETWORK if we assume that the wrong
config is already in place, but if the creation of a wrong config is not
included here, then it can only be exploited if someone else made the
exact config the attacker wants. Highly unlikely.

- "attackComplexity": "HIGH",

Since an attack here requires that you somehow make a user get a special
config file with specific properties.

- "privilegesRequired": "LOW",

Such an attack requires that the attacker can edit, create or otherwise
change files owned by the user that runs curl.

+              "confidentialityImpact": "LOW"

Because the leaked password is typically used for a limited resource,
most possibly a specific web based system/areas.